requirements.txt
openai
python-dotenv
pandas
pydantic
requests
beautifulsoup4
html2text
langchain 
langchain_ollama
tiktoken
selenium
readability-lxml
streamlit
streamlit-tags
openpyxl
groq
lxml 
html5lib
google-generativeai
webdriver-manager